RemoveBackground/requirements.txt
pillow == 10.2.0
rembg == 2.0.56